Rename WebFinger classes to make them consistent
