+++ owncloudData/src/main/java/com/owncloud/android/data/webfinger/datasources/RemoteWebFingerDatasource.kt
@@ -1,8 +1,7 @@
 /**
  * ownCloud Android client application
  *
- * @author Abel García de Prada
- * Copyright (C) 2023 ownCloud GmbH.
+ * Copyright (C) 2022 ownCloud GmbH.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2,
@@ -16,11 +15,21 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.owncloud.android.domain.webfinger.model
+package com.owncloud.android.data.webfinger.datasources
 
-enum class WebfingerRel(val uri: String) {
-    OWNCLOUD_INSTANCE("http://webfinger.owncloud/rel/server-instance"),
+import com.owncloud.android.domain.webfinger.model.WebFingerRel
 
-    // https://openid.net/specs/openid-connect-discovery-1_0.html#IssuerDiscovery
-    OIDC_ISSUER_DISCOVERY("http://openid.net/specs/connect/1.0/issuer")
+interface RemoteWebFingerDatasource {
+    fun getInstancesFromWebFinger(
+        lookupServer: String,
+        rel: WebFingerRel,
+        username: String
+    ): List<String>
+
+    fun getInstancesFromAuthenticatedWebfinger(
+        lookupServer: String,
+        rel: WebFingerRel,
+        username: String,
+        accessToken: String,
+    ): List<String>
 }
